Update patch set 1
Patch Set 1:

As far as I understand, the qseecom daemon normally first looks for it in /vendor/firmware/keymaster/. For many devices it seems to be enough to symlink /system/vendor to /vendor which they usually do for other purposes anyway.

With that said, we're also having a problem with keymaster on the m8. Have you verified that this works while simply having qseecom load it from /vendor doesn't?

Patch-set: 1