Update patch set 1
Patch Set 1:

We use bacon's qseecomd now and it definitely needs keymaster where it is after this commit (logcat and functional behavior could not be more explicit).

I wanted to avoid this and used shamu's before which worked somehow, but I don't think mixing and matching blobs is a good thing. [Blobs are the shit anyway, I hate Google so much for their licensing mess.]

Patch-set: 1