Update patch set 1
Patch Set 1:

libQSEEComAPI.so holds our required functions, which is tied to (as in "should be used with") the qseecomd binary. bacon's is just fine and properly tested in combination with H/W FDE.

I can only comment on logcat errors and functional failure without the keymaster files in "/firmware/image". As we depend on closed source files in a sensitive area (I had too many negative encounters with various crypto bits failing catastrophically), I'd prefer it to just work reliably.

Patch-set: 1