Update patch set 1
Patch Set 1:

The code for qseecomd apparently shows that it was changed to look only in /firmware/image in the Lollipop version of the qseecom blob. Why it was changed is anyone's guess.

What would happen if a user flashed a new radio image and then rebooted into CM? Wouldn't that then overwrite the /firmware partition and break the hw-backed keystore until they reflashed CM?

Patch-set: 1