Update patch set 1
Patch Set 1:

Marc: Just to be clear. I'm not trying to say that this is necessarily the "wrong" way to do it. I'm in the exact same situation (I'm pretty sure we are getting the exact same error in fact) but I just want a solutions with as few drawbacks as possible. I asked about your usage of bacon's blobs simply because I was curious and figured it might be useful information for me to know, not to question that descision.

Patch-set: 1
Reviewer: Björn Lundén <16@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0