Update patch set 1
Patch Set 1:

We are trying to force the user to use a specified modem on our device. If the user insists to mess with this, we should be free to detect this very rare condition and remount rw to copy some stuff. I didn't check if nessecary at all, but we just had to quickly mount r /firmware and see if the keymaster files are present.

On the other hand, I'm not fond of the idea to remount and symlink things back and forth every boot, just to catch a corner case and only because we're tied to a bacon's qseecom daemon.

Patch-set: 1
Reviewer: Marc vdW <6315@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0