Update patch set 1
Patch Set 1:

Oh, I didn't even notice the change in native got merged.

While I was never able to reproduce a number of > 3 for mNumPresentWithoutResyncSamples (so well with ins limit), it might still be needed as I do not know the actual hardware bug or how to repro/trigger it. For testing add this to "dumpsys SurfaceFlinger --dispsync".

This workaround should be enabled on hammerhead for peace of mind as well.

Disabling hw composing is something else or it might even be related - I just don't know.

Patch-set: 1