Update patch set 1
Patch Set 1: -Verified -Code-Review

I've noticed that the script always prefixes the "Copyright" string which I did not when doing the rebranding so I wanted to adapt it now. However, this is a pure cosmetic change. It's not a big deal for me to abandon it? :)

Patch-set: 1
Reviewer: Deleted <15228@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0